Bring in OpenSSL checkin 19821:
 Make inline assembler clang-friendly [from HEAD].

 openssl/crypto/md32_common.h     	1.45.2.1 -> 1.45.2.2
 openssl/crypto/rc5/rc5_locl.h     	1.8 -> 1.8.8.1

Approved by:	simon


Former-commit-id: d1c8a9be301a8bd44e29ede255e94e0daff08471

+++ crypto/openssl/crypto/rc5/rc5_locl.h
@@ -292,7 +292,7 @@ name|a
 parameter_list|,
 name|n
 parameter_list|)
-value|({ register unsigned int ret;	\ 					asm ("roll %%cl,%0"	\ 						: "=r"(ret)	\ 						: "c"(n),"0"(a)	\ 						: "cc");	\ 					ret;			\ 				})
+value|({ register unsigned int ret;	\ 					asm ("roll %%cl,%0"	\ 						: "=r"(ret)	\ 						: "c"(n),"0"((unsigned int)(a))	\ 						: "cc");	\ 					ret;			\ 				})
 end_define
 
 begin_define
@@ -304,7 +304,7 @@ name|a
 parameter_list|,
 name|n
 parameter_list|)
-value|({ register unsigned int ret;	\ 					asm ("rorl %%cl,%0"	\ 						: "=r"(ret)	\ 						: "c"(n),"0"(a)	\ 						: "cc");	\ 					ret;			\ 				})
+value|({ register unsigned int ret;	\ 					asm ("rorl %%cl,%0"	\ 						: "=r"(ret)	\ 						: "c"(n),"0"((unsigned int)(a))	\ 						: "cc");	\ 					ret;			\ 				})
 end_define
 
 begin_endif
